requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
databases[aiosqlite]
passlib[bcrypt]
python-jose
